Automate extensions configuration
This enhances the PHP/Extensions module with more features. Extensions
are added via add_subdirectory() according to the directory priority and
the extension dependencies listed in the custom target property. If
user passed configuration is not sufficient (for example, some
extension dependency is not enabled), module automatically enables and
configures the extensions accordingly.

This also starts utilizing CMake custom properties more since they are
useful in project configuration wide context.