Add various improvements
Update most of the functions that are found in the common historic
resolv/nsl/socket combination for the current systems:

This fixes checks for:
- gai_strerror
- socket
- socketpair
- gethostname

These are normally found in the C library on most systems except the
provided exceptions. Most exceptions include the Solaris 11.3 to 11.4
transition, illumos-based systems also have various presence but are
mostly following the 11.4 system with some left-overs. These have some
in nsl library and some in the socket library. The Windows checks have
been also updated accordingly. Haiku has most of these in the network
library. More will be updated in the future.