Fix PHP path configuration variables and install prefix
This now fixes also PHP_EXTENSION_DIR, PHP_CONFIG_FILE_PATH,
PHP_CONFIG_FILE_SCAN_DIR, and all CMAKE_INSTALL_* variables to be
properly replaced based on the given installation prefix.

The main/build-defs.h.in and main/php_config.h are still left to be done
somehow differently.

At least for the *nix systems this now works sufficiently for the time
being and will be refactored and enhanced further in the future if
possible.