Fix PHP include directory issue in scripts
There is a common issue in C projects what the includedir, a.k.a.
CMAKE_INSTALL_INCLUDEDIR variable be. It is best to be left to packaging
phase only and set to default. Projects which need to set a directory
inside the includedir should simply introduce another variable to be
able to override this if not set by default project ok. This now
intoduces a new variable `PHP_INCLUDE_PREFIX` that can do that and
adjusts all paths accordingly. Also, the custom PHP_EXPAND generator
expression hack is replaced with native genexes using PATH:ABSOLUTE_PATH
and similar. For now this should be sufficient until the next issue is
found.