Create Zend library as OBJECT library
When linking static libraries to final or transitional targets, only
those symbols that are used in the final binary will be added from the
linked library. For now this is a quick fix to get the all Zend library
symbols in the final binary and fix the shared extensions builds.
Probably a nicer and proper solution could be done in the API provided
by the C code.