Improve build types and refactor PHP_EXTENSION_DIR
For the time being this adds a new way of assembling the
PHP_EXTENSION_DIR variable that is more "correct" way and relevant as
there are different build types. Since this value is not "guessable"
and needs to be retrieved by the config script, pkg-config, or some find
module, this doesn't make any difference from the CMake-based build
system PoV. But will be probably in need of more refactorings later on.