Add config.h configuration header templates for extensions
When using Autotools and phpize, each extension has its own config.h
file generated. As this makes it very convenient to separate
configurations into separate contexts and simplify the main
configuration header template this brings similar functionality also to
CMake with possibility to further enhance it as needed.

Instead of using all-in-one main/php_config.h or main/config.w32.h
configuration header for everything in the php-src, this also seems a
more proper way forward. What is configured in some extension's scope
should only be used by the extension and not elsewhere. And when
building inside php-src, this still behaves the same as in native build
systems because all configurations are added to the main configuration
header file.