Fix extension dependencies a bit further
This now parses extensions CMakeLists.txt by only the add_dependencies()
usage. It sorts, configures extensions in the ext directory and emits
warnins or errors in a bit more intuitive manner.

Traversing of ext directories is moved to its own CMakeLists.txt for
now.

PDO* extensions dependent options simplified to regular option().

The configuration "titles" are refactored a bit due to CMake GUIs and
all sorts of log formats.

For the time being the PHP_SOURCE_DIR is indicator whether the extension
is built in the php-src or as a standalone extension.