Enable CXX and ASM optionally
CXX in PHP is at the time of writing only used in extensions (intl,
pdo_firebird as of PHP 8.4 etc.), and ASM in the Zend Engine for using
Boost assembly files.

This now resembles also native Autotools build system where CXX is
enabled optionally based on the extension that uses it.

So in case of a default build where CXX sources aren't being compiled,
CXX compiler is required only as needed.