Remove gdbm integration
GDBM (GNU dbm) library is licensed under the GPL 3 license which is not
compatible with PHP license. When linking dynamically or statically, PHP
should be licensed as GPL 3 also, otherwise it shouldn't be distributed.

This simplifies this issue and enables distributing PHP and dba
extension. The qdbm and ldbm handlers are sufficient to use the dba
extension.