Enable OBJECT and STATIC libraries for Zend and extensions
This is for testing purposes for now and to see how this behaves and if
this would work ok. This enables having extensions, as STATIC, SHARED,
MODULE, and OBJECT libraries. And Zend can be also STATIC.

STATIC libraries are linked with WHOLE_ARCHIVE feature (--whole-archive
and other linker flags) but this is more like a cutting edge feature of
linkers and probably not yet very portable solution, so Zend OBJECT
library is still a better way here to go with current state of the
php-src.

Additionally, always enabled extensions are marked as OBJECT libraries
to make it clearer.

Fixes GH-3