Enable OBJECT and STATIC libraries for Zend and extensions (#20)
This is for testing purposes for now to see how it behaves and if this
would work. This enables having extensions, as STATIC, SHARED,
MODULE, and OBJECT libraries. And Zend can be also STATIC.

STATIC libraries are linked with WHOLE_ARCHIVE feature (--whole-archive
and other linker flags) but this is more like a non-portable feature
depending on linker, so Zend OBJECT library is still a better way here
with the current state of the php-src.

Additionally, always enabled extensions are marked as OBJECT libraries
to make it clearer.

Fixes GH-3