Move config.cmake.h.in files to cmake subdirectories
As Autotools already generates config.h.in files, there is a naming
inconvenience for these templates. Projects usually solve this by naming
CMake header templates with non-standard names like config.cmake.h.in or
differently (config-cmake.h.in)... Still, the name config.h.in seems the
most convenient on the expense of residing in a subdirectory.

Another option would be to do this for extensions or SAPIs inside
CMakeLists.txt but makes CMake code a bit less readable as there is no
syntax highlighting inside such bracket or quoted arguments:

|  file(CONFIGURE OUTPUT config.h CONTENT [[
|  /* Define to 1 if the PHP extension 'bcmath' is available. */
|  #cmakedefine HAVE_BCMATH 1
|  ]])