Rename SAPI_<SAPI_NAME> variables to PHP_SAPI_<SAPI_NAME>
This is a first mitigation step to start taking care of the global state
annoyances.

All cache variables of some project should be scoped inside some
namespace. In CMake's case, at the time of writing, this can be taken
care of prefixing all variable names. For example, with 'PHP_'.

Specifically, this issue is noticed when using the CMake's
`FetchContent` module, where concerns with cache variables overrides
from one project to the other can happen as the same CMakeCache.txt
file is used for the entire build.