Rename SAPI_<SAPI_NAME> variables to PHP_SAPI_<SAPI_NAME> (#27)
This is a first mitigation step to start taking care of the global
state annoyances.

All cache variables of some project should ideally be scoped inside
some namespace. In CMake's case, at the time of writing, this can be
taken care of prefixing all variable names. For example, with 'PHP_'.

Specifically, this issue is noticed when using the CMake
`FetchContent` module, where cache variable names can collide from
one project to the other as the same `CMakeCache.txt` file is used
for the entire build.

Some more cache variables are also synced in the PHP SAPIs.