Enhance pgsql and pdo_pgsql checks
When checking for features of PostgreSQL library (libpq) there might
happen that the libpq library is of certain version (for example,
version 17), but the header file libpq-fe.h are of PostgreSQL 16 as they
are bundled together in a single postgresql-devel package. This is how
the PostgreSQL is packaged on openSUSE Linux, at the time of writing.

This is considered an upstream bug because it can cause unexpected
behavior when building software depending on libpq.

Here, a simple solution is done to check symbols with the
check_symbol_exists(), which checks if the linker sees the symbol and
if it is also declared in the header.

As of PostgreSQL version 14, there are also features macros available in
the libpq-fe.h.