Add shared SAPIs and enable position independent code globally
Changes:
* Position independent code is for now unconditionally enabled globally
* Apache PHP module output filename named to mod_php
* Added phpdbg pkg-config file for its shared library/module for the
  sake of consistency with embed and php as a whole
* Added shared embed SAPI

For the time being, this simplifies building shared apache2handler,
embed and phpdbg SAPIs that require the position independent code.

While there is this rule of a thumb that position independent code
enabled on executables slightly decreases performance, the usability
benefits outweight this issue in this case. At least until a better
solution can be integrated here.

The conditional check for CMAKE_POSITION_INDEPENDENT_CODE is removed
because now it seems to be completely redundant in any case. The
upstream issue is different because there the PIC is forcefully disabled
in configure.ac on some 32-bit machines and that's why it doesn't work.