Use target_compile_features() to set C++ standard
This is a better way to enable C++ standard for extensions in this case.
It errors out if C++ compiler doesn't support required standard and it
can be also controlled from the parent project using CMAKE_CXX_STANDARD
variable or through linking. The stricter and newer requirement is
applied in case of different configurations. Using PRIVATE scope to
limit this requirement to extension scope only.