Remove CMAKE_INCLUDE_CURRENT_DIR setting
On Windows there is win32/time.h header which causes issues with this
kind of configuration on MSVC because it seems that MSVC doesn't process
include directories the same as GCC, Clang, and others, and loads the
local file (time.h) first if the include flag is added. This is partial
fix for this issue for now.